chore: move eap outside ike/message
