+++ eap/eap_aka_prime_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	eap_message "github.com/free5gc/ike/message/eap"
+	eap_message "github.com/free5gc/ike/eap"
 )
 
 func TestEapAkaPrime(t *testing.T) {
