+++ eap/eap_expanded.go
